flatten up
