--- packages/tldraw/src/lib/shapes/geo/components/DashStyleEllipse.tsx
@@ -1,50 +1,0 @@
-import { TLGeoShape, TLShapeId, perimeterOfEllipse } from '@tldraw/editor'
-import * as React from 'react'
-import { ShapeFill, useDefaultColorTheme } from '../../shared/ShapeFill'
-import { getPerfectDashProps } from '../../shared/getPerfectDashProps'
-
-export const DashStyleEllipse = React.memo(function DashStyleEllipse({
-	w,
-	h,
-	strokeWidth: sw,
-	dash,
-	color,
-	fill,
-}: Pick<TLGeoShape['props'], 'w' | 'h' | 'dash' | 'color' | 'fill'> & {
-	strokeWidth: number
-	id: TLShapeId
-}) {
-	const theme = useDefaultColorTheme()
-	const cx = w / 2
-	const cy = h / 2
-	const rx = Math.max(0, cx)
-	const ry = Math.max(0, cy)
-
-	const perimeter = perimeterOfEllipse(rx, ry)
-
-	const { strokeDasharray, strokeDashoffset } = getPerfectDashProps(
-		perimeter < 64 ? perimeter * 2 : perimeter,
-		sw,
-		{
-			style: dash,
-			snap: 4,
-			closed: true,
-		}
-	)
-
-	const d = `M${cx - rx},${cy}a${rx},${ry},0,1,1,${rx * 2},0a${rx},${ry},0,1,1,-${rx * 2},0`
-
-	return (
-		<>
-			<ShapeFill theme={theme} d={d} color={color} fill={fill} />
-			<path
-				d={d}
-				strokeWidth={sw}
-				fill="none"
-				stroke={theme[color].solid}
-				strokeDasharray={strokeDasharray}
-				strokeDashoffset={strokeDashoffset}
-			/>
-		</>
-	)
-})
