--- packages/tldraw/src/lib/shapes/geo/components/DrawStylePolygon.tsx
@@ -1,38 +1,0 @@
-import { TLGeoShape, Vec } from '@tldraw/editor'
-import * as React from 'react'
-import { ShapeFill, useDefaultColorTheme } from '../../shared/ShapeFill'
-import { getRoundedInkyPolygonPath, getRoundedPolygonPoints } from '../geo-shape-helpers'
-
-export const DrawStylePolygon = React.memo(function DrawStylePolygon({
-	id,
-	outline,
-	lines,
-	fill,
-	color,
-	strokeWidth,
-}: Pick<TLGeoShape['props'], 'fill' | 'color'> & {
-	id: TLGeoShape['id']
-	outline: Vec[]
-	strokeWidth: number
-	lines?: Vec[][]
-}) {
-	const theme = useDefaultColorTheme()
-	const polygonPoints = getRoundedPolygonPoints(id, outline, strokeWidth / 3, strokeWidth * 2, 2)
-	let strokePathData = getRoundedInkyPolygonPath(polygonPoints)
-
-	if (lines) {
-		for (const [A, B] of lines) {
-			strokePathData += `M${A.toFixed()}L${B.toFixed()}`
-		}
-	}
-
-	const innerPolygonPoints = getRoundedPolygonPoints(id, outline, 0, strokeWidth * 2, 1)
-	const innerPathData = getRoundedInkyPolygonPath(innerPolygonPoints)
-
-	return (
-		<>
-			<ShapeFill d={innerPathData} fill={fill} color={color} theme={theme} />
-			<path d={strokePathData} stroke={theme[color].solid} strokeWidth={strokeWidth} fill="none" />
-		</>
-	)
-})
